refactor(p1): extract PhoneAccessTools.kt — device_phone_state / call_log_list / call_phone
Same factory shape. Three internal createXxxTool(context, deps) + file-private
queryCallLogs(context, limit, since).

Side cleanup: drop dead imports `android.provider.CallLog` and `android.telephony.TelephonyManager`.

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
